# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Airport Flight Data Collector API
  description: |
    Comprehensive REST and GraphQL API for accessing flight data, predictions, and analytics.
    Features include real-time flight status, ML-based delay predictions, pattern detection,
    data export, webhook integrations, and advanced analytics for SFO, YYZ, and YVR airports.
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/czhaoca/airport-flight-data-collector/issues
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.flightdata.example.com/v1
    description: Production server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login to get JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: secretpassword
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /flights:
    get:
      tags:
        - Flights
      summary: Get flight data
      description: Retrieve flight data with various filters
      parameters:
        - $ref: '#/components/parameters/Airport'
        - $ref: '#/components/parameters/FlightType'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Airline'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlightListResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /flights/{id}:
    get:
      tags:
        - Flights
      summary: Get specific flight
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Flight ID
      responses:
        '200':
          description: Flight found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlightDetailResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /predictions/batch:
    post:
      tags:
        - Predictions
      summary: Batch delay predictions
      description: Get delay predictions for multiple flights at once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchPredictionRequest'
      responses:
        '200':
          description: Batch predictions generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchPredictionResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /batch:
    post:
      tags:
        - Batch
      summary: Execute batch operations
      description: Perform multiple operations in a single request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchRequest'
      responses:
        '200':
          description: Batch operations completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Airport:
      name: airport
      in: query
      description: Filter by airport code
      schema:
        type: string
        enum: [SFO, YYZ, YVR]
    FlightType:
      name: type
      in: query
      description: Filter by flight type
      schema:
        type: string
        enum: [arrival, departure]
    StartDate:
      name: startDate
      in: query
      description: Start date (inclusive)
      schema:
        type: string
        format: date
        example: '2025-01-01'
    EndDate:
      name: endDate
      in: query
      description: End date (inclusive)
      schema:
        type: string
        format: date
        example: '2025-01-31'
    Status:
      name: status
      in: query
      description: Filter by flight status
      schema:
        type: string
        enum: [ON_TIME, DELAYED, CANCELLED, DEPARTED, ARRIVED, BOARDING]
    Airline:
      name: airline
      in: query
      description: Filter by airline code
      schema:
        type: string
        example: 'UA'
    Limit:
      name: limit
      in: query
      description: Maximum number of results
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    Offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        minimum: 0
        default: 0
    Sort:
      name: sort
      in: query
      description: Sort order
      schema:
        type: string
        enum: [scheduledTime, actualTime, flightNumber, status]
        default: scheduledTime
...
  schemas:
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: JWT refresh token
        expiresIn:
          type: integer
          description: Token expiration time in seconds
        tokenType:
          type: string
          default: Bearer
    Flight:
      type: object
      properties:
        id:
          type: string
          description: Unique flight identifier
        flightNumber:
          type: string
          example: 'UA123'
        airline:
          $ref: '#/components/schemas/Airline'
        origin:
          type: string
          example: 'LAX'
        destination:
          type: string
          example: 'SFO'
        scheduledTime:
          type: string
          format: date-time
        actualTime:
          type: string
          format: date-time
          nullable: true
        estimatedTime:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [ON_TIME, DELAYED, CANCELLED, DEPARTED, ARRIVED, BOARDING]
        gate:
          type: string
          nullable: true
        terminal:
          type: string
          nullable: true
        type:
          type: string
          enum: [arrival, departure]
        aircraft:
          type: string
          nullable: true
        delay:
          type: integer
          description: Delay in minutes
          nullable: true
    Airline:
      type: object
      properties:
        code:
          type: string
          example: 'UA'
        name:
          type: string
          example: 'United Airlines'
...
    FlightListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Flight'
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    FlightDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Flight'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
        cached:
          type: boolean
          default: false
    DelayPredictionRequest:
      type: object
      required:
        - scheduledTime
        - airline
        - origin
      properties:
        flightNumber:
          type: string
          example: 'UA123'
        airline:
          type: string
          example: 'UA'
        origin:
          type: string
          example: 'SFO'
        destination:
          type: string
          example: 'LAX'
        scheduledTime:
          type: string
          format: date-time
          example: '2025-08-02T08:00:00Z'
        aircraft:
          type: string
          example: 'B737'
        gate:
          type: string
          example: 'A12'
...
    BatchPredictionRequest:
      type: object
      required:
        - flights
      properties:
        flights:
          type: array
          items:
            $ref: '#/components/schemas/DelayPredictionRequest'
          maxItems: 100
    BatchPredictionResponse:
      type: object
      properties:
        success:
          type: boolean
        predictions:
          type: array
          items:
            type: object
            properties:
              flightId:
                type: string
              prediction:
                $ref: '#/components/schemas/DelayPredictionResponse/properties/prediction'
              error:
                type: string
                nullable: true
...
    BatchRequest:
      type: object
      required:
        - operations
      properties:
        operations:
          type: array
          items:
            type: object
            required:
              - operation
              - path
            properties:
              operation:
                type: string
                enum: [GET, POST, PUT, DELETE]
              path:
                type: string
                example: '/flights?airport=SFO&limit=10'
              body:
                type: object
                nullable: true
              id:
                type: string
                description: Optional ID to correlate request with response
          maxItems: 100
    BatchResponse:
      type: object
      properties:
        success:
          type: boolean
        results:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                nullable: true
              status:
                type: integer
              response:
                type: object
              error:
                type: string
                nullable: true
        summary:
          type: object
          properties:
            totalOperations:
              type: integer
            successfulOperations:
              type: integer
            failedOperations:
              type: integer
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: BAD_REQUEST
            message: Invalid date format
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: UNAUTHORIZED
            message: Invalid or missing authentication token
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NOT_FOUND
            message: Flight not found
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Request limit per window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Time when rate limit resets
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: RATE_LIMIT_EXCEEDED
            message: Too many requests. Please try again later.